feat: Add PWA install functionality
Integrates PWA installation capabilities into the TECCON application. This includes capturing the `beforeinstallprompt` event to enable users to install the app to their home screen and adding a new "Instalar App" button to the sidebar. The QR code links in the manifest are also updated to point to the new deployment URL.